src/backend/requirements.txt
flask
Pillow
torch
diffusers
transformers
accelerate
flask-cors
numpy
pydub
hilbertcurve
pymongo
